Move shaders to shader files
