+++ src/app/shared/services/materials/contour-material.service.ts
@@ -1,46 +1,12 @@
 import { inject, Injectable } from '@angular/core';
 import { BaseMaterialService } from './base-material.service';
 import { Color, Material, MeshNormalMaterial, ShaderMaterial } from 'three';
-import { xyzPositionVertexShader } from '../../shaders/xyzPosition.vertex-shader';
+import xyzPositionVertexShader from '../../shaders/xyzPosition.vertex-shader.glsl';
 import { SettingsService } from '../settings/settings.service';
 import { colorToVector3 } from '../../functions/color-to-vector3';
 import { LocalizeService } from '../localize.service';
 import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
-
-/**
- * Fragment shader that uses `xyzPosition` to
- * set fragment color to `contourColor` if the fragment is on
- * a multiple of `yIncrement`, or sub-increment.
- *
- * Sub-increments are 30% closer to black and are 1/3 the thickness of increments.
- *
- */
-const fragmentShader = `
-uniform float yIncrement;
-uniform vec3 contourColor;
-uniform float roundPrecision;
-uniform uint numSubIncrements;
-
-varying vec3 xyzPosition;
-
-void main()
-{
-  // primary increment
-  if (mod( xyzPosition.y, yIncrement ) < roundPrecision) {
-    gl_FragColor.rgb = contourColor;
-    gl_FragColor.a = 1.0;
-  }
-  else {
-    // sub-increments
-    for (uint i = 0u; i < numSubIncrements; ++i) {
-      if (mod( xyzPosition.y + yIncrement * float(i) / float(numSubIncrements) , yIncrement ) < roundPrecision / 3.0) {
-        gl_FragColor.rgb = mix(contourColor, vec3(0,0,0), 0.3);
-        gl_FragColor.a = 1.0;
-      }
-    }
-  }
-}
-`;
+import fragmentShader from '../../shaders/contour-material.fragment-shader.glsl';
 
 @Injectable()
 export class ContourMaterialService extends BaseMaterialService<Material[]> {
